+++ frontend/app/routes/public/status/index.tsx
@@ -18,10 +18,9 @@ import { ErrorSummaryProvider } from '~/components/error-summary-context';
 import { InlineLink } from '~/components/inline-link';
 import { InputRadios } from '~/components/input-radios';
 import { LoadingButton } from '~/components/loading-button';
-import { useFetcherSubmissionState } from '~/hooks';
+import { useFetcherSubmissionState, useHCaptcha } from '~/hooks';
 import { pageIds } from '~/page-ids';
 import { useFeature } from '~/root';
-import { useHCaptcha } from '~/utils/hcaptcha-utils';
 import { mergeMeta } from '~/utils/meta-utils';
 import type { RouteHandleData } from '~/utils/route-utils';
 import { getPathById } from '~/utils/route-utils';
