+++ frontend/app/routes/public/status/myself.tsx
@@ -19,11 +19,10 @@ import { ErrorSummary } from '~/components/error-summary';
 import { ErrorSummaryProvider } from '~/components/error-summary-context';
 import { InputPatternField } from '~/components/input-pattern-field';
 import { LoadingButton } from '~/components/loading-button';
-import { useFetcherSubmissionState } from '~/hooks';
+import { useFetcherSubmissionState, useHCaptcha } from '~/hooks';
 import { pageIds } from '~/page-ids';
 import { useFeature } from '~/root';
 import { applicationCodeInputPatternFormat, isValidCodeOrNumber } from '~/utils/application-code-utils';
-import { useHCaptcha } from '~/utils/hcaptcha-utils';
 import { mergeMeta } from '~/utils/meta-utils';
 import type { RouteHandleData } from '~/utils/route-utils';
 import { getPathById } from '~/utils/route-utils';
